+++ Katiba55.API/Entities/ProjectExecutionProgress.cs
@@ -7,7 +7,7 @@
         public string? Details { get; set; }
         public string? Notes { get; set; }
 
-        public long ProjectId { get; set; }
+        public int ProjectId { get; set; }
         public Project Project { get; set; }
     }
 }
